even more optimizations
